+++ R/delete_comments.R
@@ -3,7 +3,6 @@
 #' @param id   String. Required. id of the comment being retrieved
 #' @param \dots Additional arguments passed to \code{\link{tuber_DELETE}}.
 #' 
-#' @return 
 #' 
 #' @references \url{https://developers.google.com/youtube/v3/docs/comments/delete}
 #'