+++ R/delete_videos.R
@@ -3,7 +3,6 @@
 #' @param id   String. Required. id of the video that is to be deleted
 #' @param \dots Additional arguments passed to \code{\link{tuber_DELETE}}.
 #' 
-#' @return 
 #' 
 #' @references \url{https://developers.google.com/youtube/v3/docs/playlistItems/delete}
 #'